+++ CHANGELOG.md
@@ -8,13 +8,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Changed
+
+- The `go.opentelemetry.io/contrib/instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo` instrumentation now accepts a `WithCommandAttributeDisabled`, 
+   so the caller can specify whether to opt-out of tracing the mongo command. (#712)
+
 ## [0.19.0] - 2021-03-19
 
 ### Changed
 
 - Upgrade to v0.19.0 of `go.opentelemetry.io/otel`.
-- The `go.opentelemetry.io/contrib/instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo` instrumentation now accepts a `WithCommandLoggingDisabled`, 
-   so the caller can specify whether to opt-out of tracing the mongo command. (#712)
 
 ## [0.18.0] - 2021-03-04